1_Document Research & Theme Identification Chatbot/requirements.txt
fastapi
uvicorn
pytesseract
pillow
PyMuPDF
openai
langchain
chromadb
python-multipart
